Refactor CloudClient into workspace module
