+++ airbyte/cli/cloud/connections.py
@@ -27,7 +27,7 @@ from airbyte.cli._input import (
 )
 from airbyte.cli._output import json_output
 from airbyte.cli.cloud._cli import cloud_app
-from airbyte.cloud.client import CloudClient
+from airbyte.cloud.workspaces import CloudWorkspace
 from airbyte.exceptions import PyAirbyteInputError
 
 
@@ -46,11 +46,11 @@ def list_(
     public_api_root: ApiUrlArg = None,
 ) -> None:
     """List connections in the workspace."""
-    workspace = CloudClient.get_workspace_from_auth(
+    workspace = CloudWorkspace(
         workspace_id=workspace_id,
         client_id=client_id,
         client_secret=client_secret,
-        public_api_root=public_api_root,
+        api_root=public_api_root,
     )
     json_output([connection.get_info() for connection in workspace.list_connections()])
 
@@ -70,11 +70,11 @@ def get(
         connection_id,
         option_name="--connection-id",
     )
-    workspace = CloudClient.get_workspace_from_auth(
+    workspace = CloudWorkspace(
         workspace_id=workspace_id,
         client_id=client_id,
         client_secret=client_secret,
-        public_api_root=public_api_root,
+        api_root=public_api_root,
     )
     json_output(workspace.get_connection(resolved_connection_id).get_info())
 
@@ -95,11 +95,11 @@ def create(  # noqa: PLR0913
     public_api_root: ApiUrlArg = None,
 ) -> None:
     """Create a new connection."""
-    workspace = CloudClient.get_workspace_from_auth(
+    workspace = CloudWorkspace(
         workspace_id=workspace_id,
         client_id=client_id,
         client_secret=client_secret,
-        public_api_root=public_api_root,
+        api_root=public_api_root,
     )
     result = workspace.deploy_connection(
         connection_name=name,
@@ -127,11 +127,11 @@ def rename(
         connection_id,
         option_name="--connection-id",
     )
-    workspace = CloudClient.get_workspace_from_auth(
+    workspace = CloudWorkspace(
         workspace_id=workspace_id,
         client_id=client_id,
         client_secret=client_secret,
-        public_api_root=public_api_root,
+        api_root=public_api_root,
     )
     json_output(workspace.get_connection(resolved_connection_id).rename(name).get_info())
 
@@ -157,11 +157,11 @@ def update(
     )
     if prefix is None and selected_streams is None:
         raise PyAirbyteInputError(message="At least one update option is required.")
-    workspace = CloudClient.get_workspace_from_auth(
+    workspace = CloudWorkspace(
         workspace_id=workspace_id,
         client_id=client_id,
         client_secret=client_secret,
-        public_api_root=public_api_root,
+        api_root=public_api_root,
     )
     connection = workspace.get_connection(resolved_connection_id)
     if prefix is not None:
@@ -186,11 +186,11 @@ def enable(
         connection_id,
         option_name="--connection-id",
     )
-    connection = CloudClient.get_workspace_from_auth(
+    connection = CloudWorkspace(
         workspace_id=workspace_id,
         client_id=client_id,
         client_secret=client_secret,
-        public_api_root=public_api_root,
+        api_root=public_api_root,
     ).get_connection(resolved_connection_id)
     connection.set_enabled(enabled=True)
     json_output(connection.get_info())
@@ -211,11 +211,11 @@ def disable(
         connection_id,
         option_name="--connection-id",
     )
-    connection = CloudClient.get_workspace_from_auth(
+    connection = CloudWorkspace(
         workspace_id=workspace_id,
         client_id=client_id,
         client_secret=client_secret,
-        public_api_root=public_api_root,
+        api_root=public_api_root,
     ).get_connection(resolved_connection_id)
     connection.set_enabled(enabled=False)
     json_output(connection.get_info())
@@ -237,11 +237,11 @@ def set_(
         connection_id,
         option_name="--connection-id",
     )
-    connection = CloudClient.get_workspace_from_auth(
+    connection = CloudWorkspace(
         workspace_id=workspace_id,
         client_id=client_id,
         client_secret=client_secret,
-        public_api_root=public_api_root,
+        api_root=public_api_root,
     ).get_connection(resolved_connection_id)
     connection.set_schedule(cron_expression)
     json_output(connection.get_info())
@@ -262,11 +262,11 @@ def manual(
         connection_id,
         option_name="--connection-id",
     )
-    connection = CloudClient.get_workspace_from_auth(
+    connection = CloudWorkspace(
         workspace_id=workspace_id,
         client_id=client_id,
         client_secret=client_secret,
-        public_api_root=public_api_root,
+        api_root=public_api_root,
     ).get_connection(resolved_connection_id)
     connection.set_manual_schedule()
     json_output(connection.get_info())
@@ -288,11 +288,11 @@ def delete(
         connection_id,
         option_name="--connection-id",
     )
-    workspace = CloudClient.get_workspace_from_auth(
+    workspace = CloudWorkspace(
         workspace_id=workspace_id,
         client_id=client_id,
         client_secret=client_secret,
-        public_api_root=public_api_root,
+        api_root=public_api_root,
     )
     workspace.permanently_delete_connection(
         connection=workspace.get_connection(resolved_connection_id),
@@ -322,11 +322,11 @@ def sync(
         connection_id,
         option_name="--connection-id",
     )
-    workspace = CloudClient.get_workspace_from_auth(
+    workspace = CloudWorkspace(
         workspace_id=workspace_id,
         client_id=client_id,
         client_secret=client_secret,
-        public_api_root=public_api_root,
+        api_root=public_api_root,
     )
     result = workspace.get_connection(resolved_connection_id).run_sync(
         wait=wait,
