+++ airbyte/cli/cloud/sources.py
@@ -28,7 +28,7 @@ from airbyte.cli._input import (
 )
 from airbyte.cli._output import json_output
 from airbyte.cli.cloud._cli import cloud_app
-from airbyte.cloud.client import CloudClient
+from airbyte.cloud.workspaces import CloudWorkspace
 
 
 sources_app = _create_app(name="sources", help_text="Manage Airbyte Cloud sources.")
@@ -44,11 +44,11 @@ def list_(
     public_api_root: ApiUrlArg = None,
 ) -> None:
     """List sources in the workspace."""
-    workspace = CloudClient.get_workspace_from_auth(
+    workspace = CloudWorkspace(
         workspace_id=workspace_id,
         client_id=client_id,
         client_secret=client_secret,
-        public_api_root=public_api_root,
+        api_root=public_api_root,
     )
     json_output([source.get_info() for source in workspace.list_sources()])
 
@@ -64,11 +64,11 @@ def get(
 ) -> None:
     """Get details of a specific source."""
     resolved_source_id = resolve_entity_id(source_id_args, source_id, option_name="--source-id")
-    workspace = CloudClient.get_workspace_from_auth(
+    workspace = CloudWorkspace(
         workspace_id=workspace_id,
         client_id=client_id,
         client_secret=client_secret,
-        public_api_root=public_api_root,
+        api_root=public_api_root,
     )
     json_output(workspace.get_source(resolved_source_id).get_info())
 
@@ -92,11 +92,11 @@ def create(
     """Create a new source in the workspace."""
     config = parse_config_options(config_json=config_json, config_file=config_file)
     config["sourceType"] = source_type
-    workspace = CloudClient.get_workspace_from_auth(
+    workspace = CloudWorkspace(
         workspace_id=workspace_id,
         client_id=client_id,
         client_secret=client_secret,
-        public_api_root=public_api_root,
+        api_root=public_api_root,
     )
     json_output(workspace.deploy_source(name=name, config=config).get_info())
 
@@ -113,11 +113,11 @@ def rename(
 ) -> None:
     """Rename a source."""
     resolved_source_id = resolve_entity_id(source_id_args, source_id, option_name="--source-id")
-    workspace = CloudClient.get_workspace_from_auth(
+    workspace = CloudWorkspace(
         workspace_id=workspace_id,
         client_id=client_id,
         client_secret=client_secret,
-        public_api_root=public_api_root,
+        api_root=public_api_root,
     )
     json_output(workspace.get_source(resolved_source_id).rename(name).get_info())
 
@@ -140,11 +140,11 @@ def update(
     """Update a source configuration."""
     resolved_source_id = resolve_entity_id(source_id_args, source_id, option_name="--source-id")
     config = parse_config_options(config_json=config_json, config_file=config_file)
-    workspace = CloudClient.get_workspace_from_auth(
+    workspace = CloudWorkspace(
         workspace_id=workspace_id,
         client_id=client_id,
         client_secret=client_secret,
-        public_api_root=public_api_root,
+        api_root=public_api_root,
     )
     json_output(workspace.get_source(resolved_source_id).update_config(config).get_info())
 
@@ -161,11 +161,11 @@ def delete(
 ) -> None:
     """Delete a source."""
     resolved_source_id = resolve_entity_id(source_id_args, source_id, option_name="--source-id")
-    workspace = CloudClient.get_workspace_from_auth(
+    workspace = CloudWorkspace(
         workspace_id=workspace_id,
         client_id=client_id,
         client_secret=client_secret,
-        public_api_root=public_api_root,
+        api_root=public_api_root,
     )
     workspace.permanently_delete_source(
         source=workspace.get_source(resolved_source_id),
