+++ airbyte/cloud/workspaces.py
@@ -43,12 +43,8 @@ from typing import TYPE_CHECKING, Any, Literal, overload
 import yaml
 
 from airbyte import exceptions as exc
-from airbyte._util import api_util, text_util
+from airbyte._util import api_imports, api_util, text_util
 from airbyte._util.api_util import get_web_url_root
-from airbyte.cloud import client as cloud_client
-from airbyte.cloud.auth import (
-    resolve_cloud_workspace_id,
-)
 from airbyte.cloud.client_config import CloudClientConfig
 from airbyte.cloud.connections import CloudConnection
 from airbyte.cloud.connectors import (
@@ -56,8 +52,16 @@ from airbyte.cloud.connectors import (
     CloudSource,
     CustomCloudSourceDefinition,
 )
+from airbyte.cloud.credentials import (
+    CREDENTIALS_FILE_PATH,
+    CloudCredentials,
+    CloudLoginResult,
+    login_with_client_credentials,
+    resolve_cloud_credentials,
+)
+from airbyte.cloud.credentials import logout as remove_credentials_file
 from airbyte.destinations.base import Destination
-from airbyte.exceptions import AirbyteError
+from airbyte.exceptions import AirbyteError, AirbyteMissingResourceError
 from airbyte.secrets.base import SecretString
 
 
@@ -66,6 +70,258 @@ if TYPE_CHECKING:
 
     from airbyte._util.api_imports import JobResponse, WorkspaceResponse
     from airbyte.sources.base import Source
+
+
+@dataclass(kw_only=True)
+class CloudClient:
+    """Authenticated client for Airbyte Cloud and self-managed Airbyte APIs."""
+
+    client_id: SecretString | None = None
+    client_secret: SecretString | None = None
+    bearer_token: SecretString | None = None
+    public_api_root: str = api_util.CLOUD_API_ROOT
+    config_api_root: str | None = None
+    organization_id: str | None = None
+
+    def __post_init__(self) -> None:
+        """Wrap provided secret values."""
+        if self.client_id is not None:
+            self.client_id = SecretString(self.client_id)
+        if self.client_secret is not None:
+            self.client_secret = SecretString(self.client_secret)
+        if self.bearer_token is not None:
+            self.bearer_token = SecretString(self.bearer_token)
+
+    @classmethod
+    def from_env(
+        cls,
+        *,
+        client_id: str | SecretString | None = None,
+        client_secret: str | SecretString | None = None,
+        bearer_token: str | SecretString | None = None,
+        organization_id: str | None = None,
+        public_api_root: str | None = None,
+        config_api_root: str | None = None,
+    ) -> CloudClient:
+        """Create a client from shared environment and credentials-file resolution."""
+        credentials = resolve_cloud_credentials(
+            client_id=client_id,
+            client_secret=client_secret,
+            bearer_token=bearer_token,
+            organization_id=organization_id,
+            public_api_root=public_api_root,
+            config_api_root=config_api_root,
+        )
+        return cls.from_credentials(credentials)
+
+    @classmethod
+    def from_auth(
+        cls,
+        *,
+        organization_id: str | None = None,
+        client_id: str | SecretString | None = None,
+        client_secret: str | SecretString | None = None,
+        bearer_token: str | SecretString | None = None,
+        public_api_root: str | None = None,
+        config_api_root: str | None = None,
+        credentials_file_path: Path = CREDENTIALS_FILE_PATH,
+    ) -> CloudClient:
+        """Create a client from explicit inputs, env vars, and credentials file."""
+        credentials = resolve_cloud_credentials(
+            organization_id=organization_id,
+            client_id=client_id,
+            client_secret=client_secret,
+            bearer_token=bearer_token,
+            public_api_root=public_api_root,
+            config_api_root=config_api_root,
+            credentials_file_path=credentials_file_path,
+        )
+        return cls.from_credentials(credentials)
+
+    def login(
+        self,
+        *,
+        interactive: bool | None = None,
+        credentials_file_path: Path = CREDENTIALS_FILE_PATH,
+    ) -> CloudLoginResult:
+        """Log in to Airbyte and persist local credentials."""
+        if interactive is True:
+            # TK-TODO: Implement and verify interactive browser login before merging this PR.
+            raise NotImplementedError("Interactive Airbyte Cloud login is not implemented.")
+        if self.client_id is not None and self.client_secret is not None:
+            return login_with_client_credentials(
+                client_id=str(self.client_id),
+                client_secret=str(self.client_secret),
+                airbyte_api_root=self.public_api_root,
+                config_api_root=self.config_api_root,
+                credentials_file_path=credentials_file_path,
+            )
+        if interactive is False:
+            raise exc.PyAirbyteInputError(
+                message="Client ID and client secret are both required.",
+                guidance="Provide both client ID and client secret for non-interactive login.",
+            )
+
+        # TK-TODO: Implement and verify interactive browser login before merging this PR.
+        raise NotImplementedError("Interactive Airbyte Cloud login is not implemented.")
+
+    def logout(
+        self,
+        *,
+        credentials_file_path: Path = CREDENTIALS_FILE_PATH,
+    ) -> None:
+        """Log out by removing locally stored credentials."""
+        remove_credentials_file(credentials_file_path=credentials_file_path)
+
+    @classmethod
+    def from_credentials(cls, credentials: CloudCredentials) -> CloudClient:
+        """Create a client from resolved Cloud credentials."""
+        return cls(
+            client_id=credentials.client_id,
+            client_secret=credentials.client_secret,
+            bearer_token=credentials.bearer_token,
+            public_api_root=credentials.public_api_root,
+            config_api_root=credentials.config_api_root,
+            organization_id=credentials.organization_id,
+        )
+
+    @classmethod
+    def from_explicit_credentials(
+        cls,
+        *,
+        client_id: str | SecretString | None = None,
+        client_secret: str | SecretString | None = None,
+        bearer_token: str | SecretString | None = None,
+        public_api_root: str | None = None,
+        config_api_root: str | None = None,
+        organization_id: str | None = None,
+    ) -> CloudClient:
+        """Create a client from explicit credential values only."""
+        return cls.from_credentials(
+            CloudCredentials(
+                client_id=SecretString(client_id) if client_id else None,
+                client_secret=SecretString(client_secret) if client_secret else None,
+                bearer_token=SecretString(bearer_token) if bearer_token else None,
+                public_api_root=public_api_root or api_util.CLOUD_API_ROOT,
+                config_api_root=config_api_root,
+                organization_id=organization_id,
+            )
+        )
+
+    def get_workspace(self, workspace_id: str | None = None) -> CloudWorkspace:
+        """Create a `CloudWorkspace` using this client's credentials."""
+        if not workspace_id:
+            raise exc.PyAirbyteInputError(
+                message="Workspace ID is required.",
+                guidance="Provide a workspace ID.",
+            )
+
+        return CloudWorkspace(
+            workspace_id=workspace_id,
+            client_id=self.client_id,
+            client_secret=self.client_secret,
+            bearer_token=self.bearer_token,
+            api_root=self.public_api_root,
+            config_api_root=self.config_api_root,
+        )
+
+    def list_workspaces(
+        self,
+        name: str | None = None,
+        *,
+        name_filter: Callable[[str], bool] | None = None,
+    ) -> list[api_imports.WorkspaceResponse]:
+        """List workspaces available to this client."""
+        return api_util.list_workspaces(
+            workspace_id="",
+            api_root=self.public_api_root,
+            name=name,
+            name_filter=name_filter,
+            client_id=self.client_id,
+            client_secret=self.client_secret,
+            bearer_token=self.bearer_token,
+        )
+
+    def list_workspaces_in_organization(
+        self,
+        organization_id: str | None = None,
+        *,
+        name_contains: str | None = None,
+        max_items_limit: int | None = None,
+    ) -> list[dict[str, object]]:
+        """List workspaces in an organization using the Config API."""
+        resolved_organization_id = organization_id or self.organization_id
+        if not resolved_organization_id:
+            raise exc.PyAirbyteInputError(
+                message="Organization ID is required.",
+                guidance="Provide an organization ID.",
+            )
+
+        return api_util.list_workspaces_in_organization(
+            organization_id=resolved_organization_id,
+            api_root=self.public_api_root,
+            config_api_root=self.config_api_root,
+            client_id=self.client_id,
+            client_secret=self.client_secret,
+            bearer_token=self.bearer_token,
+            name_contains=name_contains,
+            max_items_limit=max_items_limit,
+        )
+
+    def get_organization(
+        self,
+        organization_id: str | None = None,
+        *,
+        organization_name: str | None = None,
+    ) -> CloudOrganization:
+        """Resolve an organization by ID or exact name."""
+        if organization_id and organization_name:
+            raise exc.PyAirbyteInputError(
+                message="Provide either organization ID or organization name."
+            )
+        if not organization_id and not organization_name:
+            raise exc.PyAirbyteInputError(
+                message="Organization ID or organization name is required."
+            )
+
+        organizations = api_util.list_organizations_for_user(
+            api_root=self.public_api_root,
+            client_id=self.client_id,
+            client_secret=self.client_secret,
+            bearer_token=self.bearer_token,
+        )
+        if organization_id:
+            matching_organizations = [
+                organization
+                for organization in organizations
+                if organization.organization_id == organization_id
+            ]
+        else:
+            matching_organizations = [
+                organization
+                for organization in organizations
+                if organization.organization_name == organization_name
+            ]
+
+        if not matching_organizations:
+            raise AirbyteMissingResourceError(resource_type="organization")
+        if len(matching_organizations) > 1:
+            raise exc.PyAirbyteInputError(
+                message="Organization name matches multiple organizations."
+            )
+
+        organization = matching_organizations[0]
+
+        return CloudOrganization(
+            organization_id=organization.organization_id,
+            organization_name=organization.organization_name,
+            email=organization.email,
+            api_root=self.public_api_root,
+            config_api_root=self.config_api_root,
+            client_id=self.client_id,
+            client_secret=self.client_secret,
+            bearer_token=self.bearer_token,
+        )
 
 
 class CloudOrganization:
@@ -208,7 +464,7 @@ class CloudOrganization:
         return api_util.is_account_locked(self.payment_status, self.subscription_status)
 
 
-@dataclass(kw_only=True)  # noqa: PLR0904  # Core cloud API facade.
+@dataclass(init=False, kw_only=True)  # noqa: PLR0904  # Core cloud API facade.
 class CloudWorkspace:
     """A remote workspace on the Airbyte Cloud.
 
@@ -238,25 +494,48 @@ class CloudWorkspace:
     """
 
     workspace_id: str
-    client_id: SecretString | None = None
-    client_secret: SecretString | None = None
-    api_root: str = api_util.CLOUD_API_ROOT
-    config_api_root: str | None = None
+    client_id: SecretString | None
+    client_secret: SecretString | None
+    api_root: str
+    config_api_root: str | None
     """The Config API root URL."""
-    bearer_token: SecretString | None = None
+    bearer_token: SecretString | None
 
-    # Internal credentials object (set in __post_init__, excluded from __init__)
+    # Internal credentials object (set in __init__, excluded from repr)
     _credentials: CloudClientConfig | None = field(default=None, init=False, repr=False)
 
-    def __post_init__(self) -> None:
+    def __init__(
+        self,
+        workspace_id: str | None = None,
+        client_id: str | SecretString | None = None,
+        client_secret: str | SecretString | None = None,
+        api_root: str | None = None,
+        config_api_root: str | None = None,
+        bearer_token: str | SecretString | None = None,
+    ) -> None:
         """Validate and initialize credentials."""
-        # Wrap secrets in SecretString if provided
-        if self.client_id is not None:
-            self.client_id = SecretString(self.client_id)
-        if self.client_secret is not None:
-            self.client_secret = SecretString(self.client_secret)
-        if self.bearer_token is not None:
-            self.bearer_token = SecretString(self.bearer_token)
+        self.config_api_root = config_api_root
+        credentials = resolve_cloud_credentials(
+            workspace_id=workspace_id,
+            client_id=client_id,
+            client_secret=client_secret,
+            bearer_token=bearer_token,
+            public_api_root=api_root,
+            config_api_root=self.config_api_root,
+        )
+        if not credentials.workspace_id:
+            raise exc.PyAirbyteInputError(
+                message="Workspace ID is required.",
+                guidance="Provide a workspace ID.",
+            )
+
+        client = CloudClient.from_credentials(credentials)
+        self.workspace_id = credentials.workspace_id or ""
+        self.client_id = client.client_id
+        self.client_secret = client.client_secret
+        self.bearer_token = client.bearer_token
+        self.api_root = client.public_api_root
+        self.config_api_root = client.config_api_root
 
         # Create internal CloudClientConfig object (validates mutual exclusivity)
         self._credentials = CloudClientConfig(
@@ -318,11 +597,11 @@ class CloudWorkspace:
             workspace = CloudWorkspace.from_env(workspace_id="your-workspace-id")
             ```
         """
-        client = cloud_client.CloudClient.from_env(
-            public_api_root=api_root,
+        return cls(
+            workspace_id=workspace_id,
+            api_root=api_root,
             config_api_root=config_api_root,
         )
-        return client.get_workspace(resolve_cloud_workspace_id(workspace_id))
 
     @property
     def workspace_url(self) -> str | None:
@@ -789,7 +1068,7 @@ class CloudWorkspace:
         name_filter: Callable | None = None,
     ) -> list[WorkspaceResponse]:
         """List workspaces available to the current credentials."""
-        return cloud_client.CloudClient(
+        return CloudClient(
             client_id=self.client_id,
             client_secret=self.client_secret,
             bearer_token=self.bearer_token,
